% Generated from roxygen comments for offline documentation. Do not edit by hand.
\name{AgregarDatos}
\alias{AgregarDatos}
\title{Agregar filas a una tabla existente}
\usage{
AgregarDatos(...)
}
\arguments{
  \item{df}{`data.frame` con filas a agregar.}
  \item{tabla}{Nombre de la tabla destino.}
  \item{bd}{Nombre de la base de datos.}
}
\value{Invisible `TRUE` si exitoso.}
\description{
Agregar filas a una tabla existente
}
